+++ index.d.ts
@@ -1,3 +1,7 @@
-export default function transform(css: string, options?: {
-	parseMediaQueries?: boolean;
-}): { [selector: string]: unknown; };
+export default function transform(
+  css: string,
+  options?: {
+    ignoreRule?: (selector: string) => boolean;
+    parseMediaQueries?: boolean;
+  },
+): { [selector: string]: unknown };